requirements.txt
b2
logbook
PyYAML